+++ src/router/router_test.go
@@ -1,13 +1,9 @@
 package router
 
-import (
-	"testing"
-)
-
-func TestRoutes(t *testing.T) {
-	var r = Routes()
-	t.Log(r)
-}
+// func TestRoutes(t *testing.T) {
+// 	var r = Routes()
+// 	t.Log(r)
+// }
 
 // func TestLoginHandler(t *testing.T) {
 // 	var login = LoginHandler(w, r)
